# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Staff Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  document_number: string;
  document_date: Date;
  subject: string;
  creator_id: number;
  /** @nullable */
  creator_name?: string | null;
  current_status: string;
  file_path: string;
  created_at: Date;
  updated_at: Date;
}