# lib/api-zod/src/generated/types/documentLog.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Staff Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentLog {
  id: number;
  document_id: number;
  /** @nullable */
  user_id?: number | null;
  /** @nullable */
  user_name?: string | null;
  action: string;
  /** @nullable */
  notes?: string | null;
  timestamp: Date;
}